// src/types/eden/Product.ts
import { t } from 'elysia';
export const ProductOne = t.Object({
  SequenceNumber: t.Number(),
  LastModifiedTime: t.Date(),
  Code: t.String(),
  SuppliersCode: t.Optional(t.String()),
  Supplier: t.Optional(t.String()),
  Description: t.String(),
  Comment: t.Optional(t.String()),
  Category1: t.Optional(t.String()),
  Category2: t.Optional(t.String()),
  Category3: t.Optional(t.String()),
  Category4: t.Optional(t.String()),
  SalesAcct: t.Optional(t.String()),
  COGAcct: t.Optional(t.String()),
  StockAcct: t.Optional(t.String()),
  UserNum: t.Optional(t.Number()),
  SellUnit: t.Optional(t.String()),
  SellPrice: t.Optional(t.Number()),
  Plussage: t.Optional(t.Number()),
  BuyWeight: t.Optional(t.Number()),
  BuyUnit: t.Optional(t.String()),
  CostPrice: t.Optional(t.Number()),
  ConversionFactor: t.Optional(t.Number()),
  MarginWarning: t.Optional(t.Number()),
  SellDiscount: t.Optional(t.Number()),
  SellDiscountMode: t.Optional(t.Number()),
  UserText: t.Optional(t.String()),
  StockOnHand: t.Optional(t.Number()),
  StockValue: t.Optional(t.Number()),
  MinBuildQty: t.Optional(t.Number()),
  NormalBuildQty: t.Optional(t.Number()),
  ReorderLevel: t.Optional(t.Number()),
  JobPricingMode: t.Optional(t.Number()),
  Flags: t.Optional(t.Number()),
  Colour: t.Optional(t.Number()),
  UseMultiplePrices: t.Optional(t.Boolean()),
  SellPriceB: t.Optional(t.Number()),
  SellPriceC: t.Optional(t.Number()),
  SellPriceD: t.Optional(t.Number()),
  SellPriceE: t.Optional(t.Number()),
  SellPriceF: t.Optional(t.Number()),
  QtyBreak1: t.Optional(t.Number()),
  QtyBreak2: t.Optional(t.Number()),
  QtyBreak3: t.Optional(t.Number()),
  QtyBreak4: t.Optional(t.Number()),
  QtyBrkSellPriceA1: t.Optional(t.Number()),
  QtyBrkSellPriceA2: t.Optional(t.Number()),
  QtyBrkSellPriceA3: t.Optional(t.Number()),
  QtyBrkSellPriceA4: t.Optional(t.Number()),
  QtyBrkSellPriceB1: t.Optional(t.Number()),
  QtyBrkSellPriceB2: t.Optional(t.Number()),
  QtyBrkSellPriceB3: t.Optional(t.Number()),
  QtyBrkSellPriceB4: t.Optional(t.Number()),
  Type: t.Optional(t.String()),
  Count: t.Optional(t.Number()),
  OnOrder: t.Optional(t.Number()),
  StockTakeStartQty: t.Optional(t.Number()),
  StockTakeValue: t.Optional(t.Number()),
  StockTakeNewQty: t.Optional(t.Number()),
  BarCode: t.Optional(t.String()),
  BuyPriceCurrency: t.Optional(t.String()),
  BuyPrice: t.Optional(t.Number()),
  Custom1: t.Optional(t.String()),
  Custom2: t.Optional(t.String()),
  Custom3: t.Optional(t.String()),
  Custom4: t.Optional(t.String()),
  BuyTaxCodeOverride: t.Optional(t.String()),
  SellTaxCodeOverride: t.Optional(t.String()),
  LeadTimeDays: t.Optional(t.Number()),
  Hash: t.Optional(t.Number()),
  SellWeight: t.Optional(t.Number()),
  Custom5: t.Optional(t.String()),
  Custom6: t.Optional(t.String()),
  Custom7: t.Optional(t.String()),
  Custom8: t.Optional(t.String()),
  TaggedText: t.Optional(t.String()),
});
export const ProductMany = t.Object({
  data: t.Array(ProductOne),
  pagination: t.Object({
    total: t.Number(),
    limit: t.Number(),
    offset: t.Number(),
    next: t.Number(),
    prev: t.Number(),
  })
});